Converting Sparklines to MiniChart typescript
