+++ components/MiniChart/functions/max.ts
@@ -1,0 +1,3 @@
+export function maximum(data: any) {
+	return Math.max(...data)
+}
