+++ components/MiniChart/functions/min.ts
@@ -1,0 +1,3 @@
+export function minimum(data: any) {
+	return Math.min(...data)
+}
