--- components/Sparklines/Sparklines.js
@@ -1,77 +1,0 @@
-import PropTypes from 'prop-types'
-import React, { PureComponent } from 'react'
-
-import SparklinesLine from './SparklinesLine'
-import SparklinesReferenceLine from './SparklinesReferenceLine'
-import SparklinesFinancial from './SparklinesFinancial'
-
-import dataToPoints from './dataProcessing/dataToPoints'
-
-class Sparklines extends PureComponent {
-	static propTypes = {
-		previousClose: PropTypes.number,
-		data: PropTypes.array,
-		limit: PropTypes.number,
-		width: PropTypes.number,
-		height: PropTypes.number,
-		svgWidth: PropTypes.number,
-		svgHeight: PropTypes.number,
-		preserveAspectRatio: PropTypes.string,
-		margin: PropTypes.number,
-		style: PropTypes.object,
-		min: PropTypes.number,
-		max: PropTypes.number,
-		onMouseMove: PropTypes.func
-	}
-
-	static defaultProps = {
-		data: [],
-		width: 240,
-		height: 60,
-		//Scale the graphic content of the given element non-uniformly if necessary such that the element's bounding box exactly matches the viewport rectangle.
-		preserveAspectRatio: 'none', //https://www.w3.org/TR/SVG/coords.html#PreserveAspectRatioAttribute
-		margin: 2
-	}
-
-	constructor(props) {
-		super(props)
-	}
-
-	render() {
-		const {
-			previousClose,
-			data,
-			limit,
-			width,
-			height,
-			svgWidth,
-			svgHeight,
-			preserveAspectRatio,
-			margin,
-			style,
-			max,
-			min
-		} = this.props
-
-		if (data.length === 0) return null
-		data.push(previousClose)
-		const points = dataToPoints({ previousClose, data, limit, width, height, margin, max, min })
-
-		const prevCloseCoords = points.pop()
-		data.pop()
-
-		const svgOpts = { style: style, viewBox: `0 0 ${width} ${height}`, preserveAspectRatio: preserveAspectRatio }
-		if (svgWidth > 0) svgOpts.width = svgWidth
-		if (svgHeight > 0) svgOpts.height = svgHeight
-
-		return (
-			<svg {...svgOpts}>
-				{React.Children.map(this.props.children, function (child) {
-					return React.cloneElement(child, { prevCloseCoords, data, points, width, height, margin })
-				})}
-			</svg>
-		)
-	}
-}
-
-export { Sparklines, SparklinesLine, SparklinesReferenceLine, SparklinesFinancial }
