--- components/Sparklines/dataProcessing/index.js
@@ -1,8 +1,0 @@
-export { min } from './min'
-export { max } from './min'
-export { mean } from './mean'
-export { avg } from './mean'
-export { midRange } from './midRange'
-export { median } from './median'
-export { stdev } from './stdev'
-export { variance } from './variance'
