--- components/Sparklines/dataProcessing/max.js
@@ -1,2 +1,0 @@
-// eslint-disable-next-line import/no-anonymous-default-export
-export default data => Math.max.apply(Math, data)
