--- components/Sparklines/dataProcessing/mean.js
@@ -1,2 +1,0 @@
-// eslint-disable-next-line import/no-anonymous-default-export
-export default data => data.reduce((a, b) => a + b) / data.length
