--- components/Sparklines/dataProcessing/median.js
@@ -1,2 +1,0 @@
-// eslint-disable-next-line import/no-anonymous-default-export
-export default data => data.sort((a, b) => a - b)[Math.floor(data.length / 2)]
