--- components/Sparklines/dataProcessing/midRange.js
@@ -1,5 +1,0 @@
-import min from './min'
-import max from './max'
-
-// eslint-disable-next-line import/no-anonymous-default-export
-export default data => max(data) - min(data) / 2
