--- components/Sparklines/dataProcessing/min.js
@@ -1,2 +1,0 @@
-// eslint-disable-next-line import/no-anonymous-default-export
-export default data => Math.min.apply(Math, data)
