--- components/Sparklines/testdata.ts
@@ -1,6041 +1,0 @@
-export const spy = {
-	meta: {
-		symbol: 'SPY',
-		interval: '1min',
-		currency: 'USD',
-		exchange_timezone: 'America/New_York',
-		exchange: 'NYSE',
-		type: 'ETF'
-	},
-	values: [
-		{
-			datetime: '2022-04-05 09:29:00',
-			open: '455.63000',
-			high: '455.69000',
-			low: '455.21000',
-			close: '455.24000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:28:00',
-			open: '455.60500',
-			high: '455.64000',
-			low: '455.49990',
-			close: '455.61500',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:27:00',
-			open: '455.68000',
-			high: '455.77000',
-			low: '455.56000',
-			close: '455.57000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:26:00',
-			open: '455.65000',
-			high: '455.68000',
-			low: '455.63000',
-			close: '455.68000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:25:00',
-			open: '455.65000',
-			high: '455.70000',
-			low: '455.65000',
-			close: '455.67500',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:24:00',
-			open: '455.68000',
-			high: '455.72000',
-			low: '455.68000',
-			close: '455.72000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:23:00',
-			open: '455.67500',
-			high: '455.69500',
-			low: '455.57500',
-			close: '455.69500',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:22:00',
-			open: '455.87000',
-			high: '455.87500',
-			low: '455.68990',
-			close: '455.69000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:21:00',
-			open: '455.81500',
-			high: '455.88000',
-			low: '455.80500',
-			close: '455.87000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:20:00',
-			open: '455.82000',
-			high: '455.82000',
-			low: '455.79000',
-			close: '455.81000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:19:00',
-			open: '455.84500',
-			high: '455.84500',
-			low: '455.77500',
-			close: '455.77500',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:18:00',
-			open: '455.87000',
-			high: '455.87000',
-			low: '455.81000',
-			close: '455.82500',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:17:00',
-			open: '455.86000',
-			high: '455.88000',
-			low: '455.85000',
-			close: '455.87000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:16:00',
-			open: '455.91000',
-			high: '455.92500',
-			low: '455.84000',
-			close: '455.84000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:15:00',
-			open: '455.80500',
-			high: '455.92500',
-			low: '455.80000',
-			close: '455.92500',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:14:00',
-			open: '455.73000',
-			high: '455.75000',
-			low: '455.70000',
-			close: '455.72000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:13:00',
-			open: '455.67500',
-			high: '455.72000',
-			low: '455.67000',
-			close: '455.71500',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:12:00',
-			open: '455.61500',
-			high: '455.68000',
-			low: '455.61500',
-			close: '455.67000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:11:00',
-			open: '455.71000',
-			high: '455.71000',
-			low: '455.64000',
-			close: '455.66010',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:10:00',
-			open: '455.67500',
-			high: '455.69000',
-			low: '455.64000',
-			close: '455.69000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:09:00',
-			open: '455.72000',
-			high: '455.72000',
-			low: '455.67000',
-			close: '455.69000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:08:00',
-			open: '455.72000',
-			high: '455.72000',
-			low: '455.69000',
-			close: '455.70500',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:07:00',
-			open: '455.62500',
-			high: '455.64000',
-			low: '455.60000',
-			close: '455.64000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:06:00',
-			open: '455.62000',
-			high: '455.62000',
-			low: '455.47500',
-			close: '455.60000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:05:00',
-			open: '455.73000',
-			high: '455.73000',
-			low: '455.64000',
-			close: '455.64000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:04:00',
-			open: '455.75000',
-			high: '455.76000',
-			low: '455.72000',
-			close: '455.73000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:03:00',
-			open: '455.67000',
-			high: '455.74500',
-			low: '455.67000',
-			close: '455.74500',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:02:00',
-			open: '455.60000',
-			high: '455.71000',
-			low: '455.59000',
-			close: '455.71000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:01:00',
-			open: '455.57000',
-			high: '455.66000',
-			low: '455.54000',
-			close: '455.60000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:00:00',
-			open: '455.46000',
-			high: '455.57000',
-			low: '455.43000',
-			close: '455.57000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:59:00',
-			open: '455.35990',
-			high: '455.35990',
-			low: '455.35900',
-			close: '455.35900',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:58:00',
-			open: '455.45960',
-			high: '455.46000',
-			low: '455.34990',
-			close: '455.34990',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:57:00',
-			open: '455.47000',
-			high: '455.47000',
-			low: '455.42000',
-			close: '455.44000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:56:00',
-			open: '455.55000',
-			high: '455.55000',
-			low: '455.46000',
-			close: '455.47000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:55:00',
-			open: '455.67000',
-			high: '455.69000',
-			low: '455.60000',
-			close: '455.60000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:54:00',
-			open: '455.62000',
-			high: '455.62000',
-			low: '455.57000',
-			close: '455.62000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:53:00',
-			open: '455.70000',
-			high: '455.70000',
-			low: '455.63000',
-			close: '455.64000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:52:00',
-			open: '455.74000',
-			high: '455.74000',
-			low: '455.73000',
-			close: '455.73000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:51:00',
-			open: '455.77000',
-			high: '455.77000',
-			low: '455.77000',
-			close: '455.77000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:50:00',
-			open: '455.74500',
-			high: '455.74500',
-			low: '455.74500',
-			close: '455.74500',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:49:00',
-			open: '455.78000',
-			high: '455.78000',
-			low: '455.72500',
-			close: '455.73000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:48:00',
-			open: '455.84000',
-			high: '455.85000',
-			low: '455.80000',
-			close: '455.81000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:47:00',
-			open: '455.83500',
-			high: '456.80000',
-			low: '455.83500',
-			close: '456.80000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:46:00',
-			open: '455.81000',
-			high: '455.81000',
-			low: '455.79000',
-			close: '455.79000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:45:00',
-			open: '455.81010',
-			high: '455.84000',
-			low: '455.81010',
-			close: '455.84000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:44:00',
-			open: '455.80000',
-			high: '455.80000',
-			low: '455.75900',
-			close: '455.75900',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:43:00',
-			open: '455.79000',
-			high: '455.84000',
-			low: '455.79000',
-			close: '455.80000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:42:00',
-			open: '455.72000',
-			high: '455.81000',
-			low: '455.72000',
-			close: '455.77000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:41:00',
-			open: '455.86000',
-			high: '455.89000',
-			low: '455.86000',
-			close: '455.89000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:40:00',
-			open: '455.84000',
-			high: '455.89000',
-			low: '455.80000',
-			close: '455.89000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:39:00',
-			open: '455.71000',
-			high: '455.80000',
-			low: '455.66050',
-			close: '455.77000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:38:00',
-			open: '455.77000',
-			high: '455.77000',
-			low: '455.77000',
-			close: '455.77000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:37:00',
-			open: '455.82000',
-			high: '455.82000',
-			low: '455.80000',
-			close: '455.80000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:36:00',
-			open: '455.80000',
-			high: '455.84000',
-			low: '455.80000',
-			close: '455.84000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:35:00',
-			open: '455.85000',
-			high: '455.85000',
-			low: '455.85000',
-			close: '455.85000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:34:00',
-			open: '455.77000',
-			high: '455.82000',
-			low: '455.77000',
-			close: '455.82000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:33:00',
-			open: '455.80000',
-			high: '455.80000',
-			low: '455.77500',
-			close: '455.77500',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:32:00',
-			open: '455.80000',
-			high: '455.80000',
-			low: '455.77000',
-			close: '455.77000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:31:00',
-			open: '455.75000',
-			high: '455.75000',
-			low: '455.72000',
-			close: '455.72000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:30:00',
-			open: '455.67000',
-			high: '455.75000',
-			low: '455.67000',
-			close: '455.75000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:29:00',
-			open: '455.63000',
-			high: '455.66000',
-			low: '455.63000',
-			close: '455.65000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:28:00',
-			open: '455.67000',
-			high: '455.69000',
-			low: '455.67000',
-			close: '455.69000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:27:00',
-			open: '455.64000',
-			high: '455.67000',
-			low: '455.63000',
-			close: '455.66000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:26:00',
-			open: '455.60000',
-			high: '455.65010',
-			low: '455.60000',
-			close: '455.65010',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:25:00',
-			open: '455.56000',
-			high: '455.59000',
-			low: '455.56000',
-			close: '455.59000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:24:00',
-			open: '455.80000',
-			high: '455.80000',
-			low: '455.60000',
-			close: '455.60000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:23:00',
-			open: '455.87000',
-			high: '455.92000',
-			low: '455.86000',
-			close: '455.86000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:22:00',
-			open: '455.81000',
-			high: '455.87000',
-			low: '455.81000',
-			close: '455.87000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:21:00',
-			open: '455.85000',
-			high: '455.85000',
-			low: '455.83000',
-			close: '455.83000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:20:00',
-			open: '455.80000',
-			high: '455.80000',
-			low: '455.80000',
-			close: '455.80000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:19:00',
-			open: '455.89000',
-			high: '455.89000',
-			low: '455.89000',
-			close: '455.89000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:18:00',
-			open: '455.89000',
-			high: '455.89000',
-			low: '455.82000',
-			close: '455.82000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:17:00',
-			open: '455.77000',
-			high: '455.83000',
-			low: '455.77000',
-			close: '455.83000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:16:00',
-			open: '455.83000',
-			high: '455.83000',
-			low: '455.80000',
-			close: '455.80000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:15:00',
-			open: '455.92000',
-			high: '455.92000',
-			low: '455.90000',
-			close: '455.90000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:14:00',
-			open: '456.00000',
-			high: '456.00000',
-			low: '455.94000',
-			close: '455.95000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:12:00',
-			open: '456.04000',
-			high: '456.04000',
-			low: '456.04000',
-			close: '456.04000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:11:00',
-			open: '455.97010',
-			high: '455.97010',
-			low: '455.97010',
-			close: '455.97010',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:10:00',
-			open: '455.98000',
-			high: '455.99000',
-			low: '455.98000',
-			close: '455.99000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:09:00',
-			open: '456.00000',
-			high: '456.00000',
-			low: '455.98000',
-			close: '455.98000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:08:00',
-			open: '455.95010',
-			high: '456.00000',
-			low: '455.95010',
-			close: '456.00000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:07:00',
-			open: '455.95000',
-			high: '455.95000',
-			low: '455.95000',
-			close: '455.95000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:06:00',
-			open: '456.10000',
-			high: '456.10000',
-			low: '455.98000',
-			close: '455.98000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:05:00',
-			open: '456.22000',
-			high: '456.22000',
-			low: '456.12000',
-			close: '456.12000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:04:00',
-			open: '456.19010',
-			high: '456.23000',
-			low: '456.19010',
-			close: '456.22000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:03:00',
-			open: '456.10000',
-			high: '456.17000',
-			low: '456.10000',
-			close: '456.17000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:02:00',
-			open: '456.08000',
-			high: '456.13000',
-			low: '456.08000',
-			close: '456.13000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:01:00',
-			open: '455.97000',
-			high: '456.11500',
-			low: '455.61000',
-			close: '456.04000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:00:00',
-			open: '455.50000',
-			high: '456.12990',
-			low: '455.45000',
-			close: '456.00000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:57:00',
-			open: '456.00000',
-			high: '456.00000',
-			low: '456.00000',
-			close: '456.00000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:56:00',
-			open: '456.04000',
-			high: '456.04000',
-			low: '456.04000',
-			close: '456.04000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:54:00',
-			open: '456.03000',
-			high: '456.03000',
-			low: '455.99000',
-			close: '455.99000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:53:00',
-			open: '455.97000',
-			high: '455.97000',
-			low: '455.97000',
-			close: '455.97000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:52:00',
-			open: '456.00000',
-			high: '456.00000',
-			low: '456.00000',
-			close: '456.00000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:50:00',
-			open: '456.05000',
-			high: '456.05000',
-			low: '456.05000',
-			close: '456.05000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:48:00',
-			open: '456.11000',
-			high: '456.11000',
-			low: '456.11000',
-			close: '456.11000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:47:00',
-			open: '456.08000',
-			high: '456.10000',
-			low: '456.08000',
-			close: '456.10000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:46:00',
-			open: '456.10000',
-			high: '456.10000',
-			low: '456.08000',
-			close: '456.08000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:45:00',
-			open: '456.10000',
-			high: '456.12000',
-			low: '456.09000',
-			close: '456.09000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:43:00',
-			open: '456.10000',
-			high: '456.10000',
-			low: '456.10000',
-			close: '456.10000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:42:00',
-			open: '455.95000',
-			high: '456.08000',
-			low: '455.94000',
-			close: '456.08000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:41:00',
-			open: '455.94000',
-			high: '455.97000',
-			low: '455.92000',
-			close: '455.92000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:40:00',
-			open: '455.97000',
-			high: '455.97000',
-			low: '455.97000',
-			close: '455.97000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:38:00',
-			open: '455.90000',
-			high: '455.90000',
-			low: '455.90000',
-			close: '455.90000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:36:00',
-			open: '455.88900',
-			high: '455.88900',
-			low: '455.88900',
-			close: '455.88900',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:35:00',
-			open: '455.76000',
-			high: '455.76000',
-			low: '455.76000',
-			close: '455.76000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:34:00',
-			open: '455.75000',
-			high: '455.75000',
-			low: '455.75000',
-			close: '455.75000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:33:00',
-			open: '455.81000',
-			high: '455.81000',
-			low: '455.75000',
-			close: '455.75000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:32:00',
-			open: '455.87000',
-			high: '455.87000',
-			low: '455.87000',
-			close: '455.87000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:31:00',
-			open: '455.80000',
-			high: '455.81000',
-			low: '455.80000',
-			close: '455.80000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:30:00',
-			open: '455.80000',
-			high: '455.85000',
-			low: '455.80000',
-			close: '455.85000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:26:00',
-			open: '455.72000',
-			high: '455.72000',
-			low: '455.72000',
-			close: '455.72000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:25:00',
-			open: '455.58000',
-			high: '455.58000',
-			low: '455.58000',
-			close: '455.58000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:24:00',
-			open: '455.58000',
-			high: '455.58000',
-			low: '455.58000',
-			close: '455.58000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:23:00',
-			open: '455.64000',
-			high: '455.64000',
-			low: '455.60000',
-			close: '455.60000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:22:00',
-			open: '455.76000',
-			high: '455.76000',
-			low: '455.76000',
-			close: '455.76000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:20:00',
-			open: '455.64000',
-			high: '455.64000',
-			low: '455.64000',
-			close: '455.64000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:19:00',
-			open: '455.63000',
-			high: '455.64000',
-			low: '455.63000',
-			close: '455.64000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:14:00',
-			open: '455.46000',
-			high: '455.51000',
-			low: '455.44000',
-			close: '455.47000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:13:00',
-			open: '455.50000',
-			high: '455.50000',
-			low: '455.44000',
-			close: '455.46000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:12:00',
-			open: '455.51000',
-			high: '455.51000',
-			low: '455.50000',
-			close: '455.50000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:10:00',
-			open: '455.56000',
-			high: '455.57900',
-			low: '455.56000',
-			close: '455.57900',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:08:00',
-			open: '455.61000',
-			high: '455.63000',
-			low: '455.60000',
-			close: '455.63000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:07:00',
-			open: '455.63000',
-			high: '455.63000',
-			low: '455.59000',
-			close: '455.59000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:06:00',
-			open: '455.70000',
-			high: '455.72000',
-			low: '455.64000',
-			close: '455.66000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:05:00',
-			open: '455.74000',
-			high: '455.74000',
-			low: '455.73000',
-			close: '455.74000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:04:00',
-			open: '455.71000',
-			high: '455.81000',
-			low: '455.68000',
-			close: '455.81000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:02:00',
-			open: '455.75000',
-			high: '455.75000',
-			low: '455.75000',
-			close: '455.75000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:01:00',
-			open: '455.81000',
-			high: '455.81000',
-			low: '455.74900',
-			close: '455.74900',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:00:00',
-			open: '455.74000',
-			high: '455.82900',
-			low: '455.74000',
-			close: '455.82900',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:59:00',
-			open: '455.79000',
-			high: '455.80000',
-			low: '455.79000',
-			close: '455.80000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:58:00',
-			open: '455.79000',
-			high: '455.79000',
-			low: '455.79000',
-			close: '455.79000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:57:00',
-			open: '455.73000',
-			high: '455.73000',
-			low: '455.73000',
-			close: '455.73000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:55:00',
-			open: '455.74000',
-			high: '455.74000',
-			low: '455.74000',
-			close: '455.74000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:54:00',
-			open: '455.81000',
-			high: '455.81000',
-			low: '455.81000',
-			close: '455.81000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:51:00',
-			open: '455.69000',
-			high: '455.71000',
-			low: '455.63000',
-			close: '455.70000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:50:00',
-			open: '455.61000',
-			high: '455.61000',
-			low: '455.61000',
-			close: '455.61000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:49:00',
-			open: '455.59000',
-			high: '455.59000',
-			low: '455.59000',
-			close: '455.59000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:48:00',
-			open: '455.52000',
-			high: '455.52000',
-			low: '455.52000',
-			close: '455.52000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:46:00',
-			open: '455.58000',
-			high: '455.58000',
-			low: '455.58000',
-			close: '455.58000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:45:00',
-			open: '455.51000',
-			high: '455.51000',
-			low: '455.47000',
-			close: '455.47000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:44:00',
-			open: '455.56000',
-			high: '455.56000',
-			low: '455.56000',
-			close: '455.56000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:41:00',
-			open: '455.48000',
-			high: '455.48000',
-			low: '455.48000',
-			close: '455.48000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:40:00',
-			open: '455.47000',
-			high: '455.47000',
-			low: '455.47000',
-			close: '455.47000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:39:00',
-			open: '455.40000',
-			high: '455.40000',
-			low: '455.35000',
-			close: '455.35000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:38:00',
-			open: '455.45000',
-			high: '455.45000',
-			low: '455.45000',
-			close: '455.45000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:37:00',
-			open: '455.33000',
-			high: '455.44000',
-			low: '455.33000',
-			close: '455.44000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:35:00',
-			open: '455.35000',
-			high: '455.35000',
-			low: '455.33000',
-			close: '455.33000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:34:00',
-			open: '455.45000',
-			high: '455.45000',
-			low: '455.40000',
-			close: '455.40000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:33:00',
-			open: '455.52000',
-			high: '455.52000',
-			low: '455.52000',
-			close: '455.52000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:32:00',
-			open: '455.52000',
-			high: '455.52000',
-			low: '455.52000',
-			close: '455.52000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:31:00',
-			open: '455.65000',
-			high: '455.65000',
-			low: '455.65000',
-			close: '455.65000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:30:00',
-			open: '455.52000',
-			high: '455.65000',
-			low: '455.52000',
-			close: '455.65000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:26:00',
-			open: '455.58000',
-			high: '455.58000',
-			low: '455.44000',
-			close: '455.44000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:24:00',
-			open: '455.41000',
-			high: '455.41000',
-			low: '455.41000',
-			close: '455.41000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:23:00',
-			open: '455.50000',
-			high: '455.53000',
-			low: '455.50000',
-			close: '455.53000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:22:00',
-			open: '455.61000',
-			high: '455.61000',
-			low: '455.61000',
-			close: '455.61000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:19:00',
-			open: '455.74000',
-			high: '455.74000',
-			low: '455.74000',
-			close: '455.74000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:17:00',
-			open: '455.74000',
-			high: '455.74000',
-			low: '455.74000',
-			close: '455.74000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:16:00',
-			open: '455.75000',
-			high: '455.80000',
-			low: '455.74000',
-			close: '455.80000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:15:00',
-			open: '455.63000',
-			high: '455.75000',
-			low: '455.63000',
-			close: '455.75000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:14:00',
-			open: '455.61000',
-			high: '455.69000',
-			low: '455.61000',
-			close: '455.69000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:12:00',
-			open: '455.53000',
-			high: '455.53000',
-			low: '455.53000',
-			close: '455.53000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:09:00',
-			open: '455.98000',
-			high: '455.98000',
-			low: '455.60000',
-			close: '455.61000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:08:00',
-			open: '456.10000',
-			high: '456.10000',
-			low: '456.10000',
-			close: '456.10000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:07:00',
-			open: '455.99000',
-			high: '456.06000',
-			low: '455.99000',
-			close: '456.06000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:03:00',
-			open: '456.05000',
-			high: '456.06000',
-			low: '456.01000',
-			close: '456.06000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:02:00',
-			open: '455.95000',
-			high: '456.00000',
-			low: '455.95000',
-			close: '456.00000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:01:00',
-			open: '455.86000',
-			high: '455.86000',
-			low: '455.86000',
-			close: '455.86000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:00:00',
-			open: '455.84000',
-			high: '455.84000',
-			low: '455.81000',
-			close: '455.83000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:59:00',
-			open: '455.85000',
-			high: '455.88000',
-			low: '455.80000',
-			close: '455.88000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:58:00',
-			open: '455.86000',
-			high: '455.86000',
-			low: '455.86000',
-			close: '455.86000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:57:00',
-			open: '455.89000',
-			high: '455.89000',
-			low: '455.82000',
-			close: '455.88000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:56:00',
-			open: '455.90000',
-			high: '455.99000',
-			low: '455.89000',
-			close: '455.99000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:55:00',
-			open: '455.97000',
-			high: '455.97000',
-			low: '455.89000',
-			close: '455.89000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:53:00',
-			open: '455.95000',
-			high: '455.95000',
-			low: '455.95000',
-			close: '455.95000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:52:00',
-			open: '455.98000',
-			high: '455.98000',
-			low: '455.98000',
-			close: '455.98000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:51:00',
-			open: '455.94000',
-			high: '455.95000',
-			low: '455.83000',
-			close: '455.83000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:50:00',
-			open: '455.80000',
-			high: '455.88000',
-			low: '455.79000',
-			close: '455.88000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:49:00',
-			open: '455.83000',
-			high: '455.83000',
-			low: '455.83000',
-			close: '455.83000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:48:00',
-			open: '455.95000',
-			high: '455.95000',
-			low: '455.88000',
-			close: '455.88000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:47:00',
-			open: '455.93000',
-			high: '455.95000',
-			low: '455.93000',
-			close: '455.95000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:46:00',
-			open: '455.84000',
-			high: '455.91000',
-			low: '455.84000',
-			close: '455.90000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:45:00',
-			open: '456.02000',
-			high: '456.02000',
-			low: '456.02000',
-			close: '456.02000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:44:00',
-			open: '456.12000',
-			high: '456.12000',
-			low: '456.12000',
-			close: '456.12000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:43:00',
-			open: '456.05000',
-			high: '456.13000',
-			low: '456.05000',
-			close: '456.13000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:41:00',
-			open: '456.10000',
-			high: '456.10000',
-			low: '456.02000',
-			close: '456.02000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:38:00',
-			open: '456.19000',
-			high: '456.19000',
-			low: '456.17000',
-			close: '456.17000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:37:00',
-			open: '456.35000',
-			high: '456.36000',
-			low: '456.35000',
-			close: '456.36000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:31:00',
-			open: '456.41000',
-			high: '456.41000',
-			low: '456.41000',
-			close: '456.41000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:30:00',
-			open: '456.40000',
-			high: '456.47000',
-			low: '456.40000',
-			close: '456.47000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:29:00',
-			open: '456.40000',
-			high: '456.40000',
-			low: '456.40000',
-			close: '456.40000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:28:00',
-			open: '456.38000',
-			high: '456.40000',
-			low: '456.35000',
-			close: '456.35000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:27:00',
-			open: '456.49000',
-			high: '456.49000',
-			low: '456.38000',
-			close: '456.38000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:26:00',
-			open: '456.47000',
-			high: '456.47000',
-			low: '456.47000',
-			close: '456.47000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:25:00',
-			open: '456.56000',
-			high: '456.56000',
-			low: '456.56000',
-			close: '456.56000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:24:00',
-			open: '456.60000',
-			high: '456.61000',
-			low: '456.47000',
-			close: '456.47000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:23:00',
-			open: '456.68000',
-			high: '456.68000',
-			low: '456.61000',
-			close: '456.61000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:21:00',
-			open: '456.68000',
-			high: '456.68000',
-			low: '456.68000',
-			close: '456.68000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:20:00',
-			open: '456.66000',
-			high: '456.70000',
-			low: '456.66000',
-			close: '456.70000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:19:00',
-			open: '456.67000',
-			high: '456.67000',
-			low: '456.67000',
-			close: '456.67000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:18:00',
-			open: '456.67000',
-			high: '456.67000',
-			low: '456.67000',
-			close: '456.67000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:17:00',
-			open: '456.65000',
-			high: '456.65000',
-			low: '456.63000',
-			close: '456.63000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:16:00',
-			open: '456.74000',
-			high: '456.74000',
-			low: '456.66000',
-			close: '456.66000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:15:00',
-			open: '456.75000',
-			high: '456.76000',
-			low: '456.75000',
-			close: '456.76000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:14:00',
-			open: '456.75000',
-			high: '456.75000',
-			low: '456.75000',
-			close: '456.75000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:13:00',
-			open: '456.69000',
-			high: '456.69000',
-			low: '456.69000',
-			close: '456.69000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:12:00',
-			open: '456.73000',
-			high: '456.77000',
-			low: '456.73000',
-			close: '456.76000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:11:00',
-			open: '456.74000',
-			high: '456.74000',
-			low: '456.70000',
-			close: '456.70000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:09:00',
-			open: '456.71000',
-			high: '456.71000',
-			low: '456.71000',
-			close: '456.71000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:08:00',
-			open: '456.73000',
-			high: '456.75000',
-			low: '456.66000',
-			close: '456.75000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:07:00',
-			open: '456.74000',
-			high: '456.74000',
-			low: '456.72000',
-			close: '456.72000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:06:00',
-			open: '456.73000',
-			high: '456.74000',
-			low: '456.73000',
-			close: '456.74000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:05:00',
-			open: '456.69000',
-			high: '456.74000',
-			low: '456.67000',
-			close: '456.67000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:03:00',
-			open: '456.64000',
-			high: '456.64000',
-			low: '456.55000',
-			close: '456.61000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:02:00',
-			open: '456.68000',
-			high: '456.68000',
-			low: '456.68000',
-			close: '456.68000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:00:00',
-			open: '456.63000',
-			high: '456.63000',
-			low: '456.61000',
-			close: '456.61000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:59:00',
-			open: '456.63000',
-			high: '456.63000',
-			low: '456.63000',
-			close: '456.63000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:58:00',
-			open: '456.55000',
-			high: '456.65000',
-			low: '456.55000',
-			close: '456.65000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:57:00',
-			open: '456.74000',
-			high: '456.74000',
-			low: '456.64000',
-			close: '456.64000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:56:00',
-			open: '456.81000',
-			high: '456.81000',
-			low: '456.75000',
-			close: '456.75000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:55:00',
-			open: '456.78000',
-			high: '456.79000',
-			low: '456.78000',
-			close: '456.79000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:54:00',
-			open: '456.81000',
-			high: '456.81000',
-			low: '456.78000',
-			close: '456.78000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:53:00',
-			open: '456.84000',
-			high: '456.84000',
-			low: '456.84000',
-			close: '456.84000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:51:00',
-			open: '456.83000',
-			high: '456.83000',
-			low: '456.83000',
-			close: '456.83000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:50:00',
-			open: '456.80000',
-			high: '456.80000',
-			low: '456.77000',
-			close: '456.78000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:49:00',
-			open: '456.84000',
-			high: '456.90000',
-			low: '456.83000',
-			close: '456.83000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:48:00',
-			open: '456.88000',
-			high: '456.88000',
-			low: '456.88000',
-			close: '456.88000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:45:00',
-			open: '457.01000',
-			high: '457.01000',
-			low: '456.92000',
-			close: '456.92000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:44:00',
-			open: '457.00000',
-			high: '457.03000',
-			low: '457.00000',
-			close: '457.03000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:43:00',
-			open: '457.05000',
-			high: '457.05000',
-			low: '457.05000',
-			close: '457.05000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:42:00',
-			open: '457.04000',
-			high: '457.10000',
-			low: '457.04000',
-			close: '457.04000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:41:00',
-			open: '457.08000',
-			high: '457.08000',
-			low: '457.04000',
-			close: '457.04000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:39:00',
-			open: '457.00000',
-			high: '457.01000',
-			low: '456.98000',
-			close: '457.01000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:38:00',
-			open: '456.99000',
-			high: '457.07000',
-			low: '456.99000',
-			close: '457.03000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:37:00',
-			open: '457.00000',
-			high: '457.00000',
-			low: '457.00000',
-			close: '457.00000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:36:00',
-			open: '456.98000',
-			high: '456.98000',
-			low: '456.90000',
-			close: '456.90000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:35:00',
-			open: '456.99000',
-			high: '456.99000',
-			low: '456.98000',
-			close: '456.98000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:34:00',
-			open: '456.91000',
-			high: '456.98000',
-			low: '456.88000',
-			close: '456.94000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:33:00',
-			open: '456.95000',
-			high: '456.95000',
-			low: '456.80000',
-			close: '456.85000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:32:00',
-			open: '457.00000',
-			high: '457.00000',
-			low: '457.00000',
-			close: '457.00000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:31:00',
-			open: '456.97000',
-			high: '456.97000',
-			low: '456.96000',
-			close: '456.97000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:30:00',
-			open: '456.96000',
-			high: '456.96000',
-			low: '456.96000',
-			close: '456.96000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:28:00',
-			open: '457.06000',
-			high: '457.06000',
-			low: '457.00000',
-			close: '457.00000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:27:00',
-			open: '457.07000',
-			high: '457.07000',
-			low: '457.05000',
-			close: '457.05000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:26:00',
-			open: '457.12000',
-			high: '457.12000',
-			low: '457.12000',
-			close: '457.12000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:24:00',
-			open: '457.24000',
-			high: '457.24000',
-			low: '457.24000',
-			close: '457.24000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:21:00',
-			open: '457.27000',
-			high: '457.27000',
-			low: '457.27000',
-			close: '457.27000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:20:00',
-			open: '457.33000',
-			high: '457.33000',
-			low: '457.33000',
-			close: '457.33000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:19:00',
-			open: '457.29000',
-			high: '457.29000',
-			low: '457.29000',
-			close: '457.29000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:18:00',
-			open: '457.29000',
-			high: '457.29000',
-			low: '457.29000',
-			close: '457.29000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:17:00',
-			open: '457.33000',
-			high: '457.33000',
-			low: '457.33000',
-			close: '457.33000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:16:00',
-			open: '457.36000',
-			high: '457.36000',
-			low: '457.36000',
-			close: '457.36000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:15:00',
-			open: '457.43000',
-			high: '457.43000',
-			low: '457.41000',
-			close: '457.41000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:14:00',
-			open: '457.29000',
-			high: '457.36000',
-			low: '457.29000',
-			close: '457.36000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:13:00',
-			open: '457.36000',
-			high: '457.36000',
-			low: '457.36000',
-			close: '457.36000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:12:00',
-			open: '457.25000',
-			high: '457.36000',
-			low: '457.25000',
-			close: '457.36000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:11:00',
-			open: '457.27000',
-			high: '457.27000',
-			low: '457.27000',
-			close: '457.27000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:10:00',
-			open: '457.17000',
-			high: '457.17000',
-			low: '457.17000',
-			close: '457.17000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:05:00',
-			open: '457.06000',
-			high: '457.07000',
-			low: '457.06000',
-			close: '457.07000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:04:00',
-			open: '457.22000',
-			high: '457.23000',
-			low: '457.11000',
-			close: '457.11000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:02:00',
-			open: '457.21000',
-			high: '457.21000',
-			low: '457.21000',
-			close: '457.21000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:01:00',
-			open: '457.29000',
-			high: '457.29000',
-			low: '457.29000',
-			close: '457.29000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:00:00',
-			open: '457.30000',
-			high: '457.30000',
-			low: '457.29000',
-			close: '457.29000',
-			volume: '0'
-		}
-	],
-	status: 'ok'
-}
-
-export const qqq = {
-	meta: {
-		symbol: 'QQQ',
-		interval: '1min',
-		currency: 'USD',
-		exchange_timezone: 'America/New_York',
-		exchange: 'NASDAQ',
-		type: 'ETF'
-	},
-	values: [
-		{
-			datetime: '2022-04-05 09:29:00',
-			open: '367.83000',
-			high: '367.86500',
-			low: '367.71000',
-			close: '367.77000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:28:00',
-			open: '367.92000',
-			high: '368.03000',
-			low: '367.78000',
-			close: '367.78000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:27:00',
-			open: '368.06500',
-			high: '368.16000',
-			low: '367.82000',
-			close: '367.92000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:26:00',
-			open: '368.03000',
-			high: '368.06000',
-			low: '367.98000',
-			close: '368.06000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:25:00',
-			open: '368.13000',
-			high: '368.18000',
-			low: '368.02900',
-			close: '368.02900',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:24:00',
-			open: '368.01100',
-			high: '368.09000',
-			low: '368.01100',
-			close: '368.09000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:23:00',
-			open: '368.02000',
-			high: '368.02000',
-			low: '367.92000',
-			close: '368.02000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:22:00',
-			open: '368.27000',
-			high: '368.27000',
-			low: '368.04000',
-			close: '368.04000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:21:00',
-			open: '368.20000',
-			high: '368.24000',
-			low: '368.19500',
-			close: '368.23000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:20:00',
-			open: '368.17000',
-			high: '368.20000',
-			low: '368.16000',
-			close: '368.20000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:19:00',
-			open: '368.27000',
-			high: '368.27000',
-			low: '368.18000',
-			close: '368.18000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:18:00',
-			open: '368.26000',
-			high: '368.27000',
-			low: '368.22000',
-			close: '368.22000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:17:00',
-			open: '368.26000',
-			high: '368.28000',
-			low: '368.21500',
-			close: '368.28000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:16:00',
-			open: '368.32000',
-			high: '368.34000',
-			low: '368.23900',
-			close: '368.24000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:15:00',
-			open: '368.18000',
-			high: '368.32000',
-			low: '368.18000',
-			close: '368.32000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:14:00',
-			open: '368.16000',
-			high: '368.16000',
-			low: '368.11000',
-			close: '368.14040',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:13:00',
-			open: '368.07000',
-			high: '368.14000',
-			low: '368.07000',
-			close: '368.14000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:12:00',
-			open: '368.02000',
-			high: '368.07000',
-			low: '368.00000',
-			close: '368.07000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:11:00',
-			open: '368.05500',
-			high: '368.07000',
-			low: '368.00000',
-			close: '368.04000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:10:00',
-			open: '368.03000',
-			high: '368.10000',
-			low: '368.01000',
-			close: '368.10000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:09:00',
-			open: '368.07000',
-			high: '368.07000',
-			low: '368.03000',
-			close: '368.07000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:08:00',
-			open: '368.00000',
-			high: '368.08000',
-			low: '368.00000',
-			close: '368.05500',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:07:00',
-			open: '367.96000',
-			high: '367.98000',
-			low: '367.89000',
-			close: '367.98000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:06:00',
-			open: '368.04000',
-			high: '368.04000',
-			low: '367.81000',
-			close: '367.94000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:05:00',
-			open: '368.12000',
-			high: '368.13000',
-			low: '368.03900',
-			close: '368.03900',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:04:00',
-			open: '368.17000',
-			high: '368.19000',
-			low: '368.14100',
-			close: '368.15000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:03:00',
-			open: '368.13000',
-			high: '368.17000',
-			low: '368.11000',
-			close: '368.14000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:02:00',
-			open: '368.06000',
-			high: '368.14000',
-			low: '368.04000',
-			close: '368.13000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:00:00',
-			open: '367.84000',
-			high: '367.88000',
-			low: '367.82000',
-			close: '367.87000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:59:00',
-			open: '367.74000',
-			high: '367.82500',
-			low: '367.74000',
-			close: '367.82000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:58:00',
-			open: '367.83000',
-			high: '367.84000',
-			low: '367.74000',
-			close: '367.75000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:57:00',
-			open: '367.80000',
-			high: '367.86000',
-			low: '367.77000',
-			close: '367.83000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:56:00',
-			open: '368.05000',
-			high: '368.05000',
-			low: '367.79000',
-			close: '367.80000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:55:00',
-			open: '368.08000',
-			high: '368.10000',
-			low: '368.04000',
-			close: '368.04000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:54:00',
-			open: '368.04000',
-			high: '368.06000',
-			low: '367.97000',
-			close: '368.06000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:53:00',
-			open: '368.07000',
-			high: '368.07000',
-			low: '368.02000',
-			close: '368.04000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:52:00',
-			open: '368.13000',
-			high: '368.13000',
-			low: '368.09000',
-			close: '368.11000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:51:00',
-			open: '368.18000',
-			high: '368.18000',
-			low: '368.12000',
-			close: '368.15000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:50:00',
-			open: '368.13000',
-			high: '368.13000',
-			low: '368.10000',
-			close: '368.13000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:49:00',
-			open: '368.11000',
-			high: '368.13000',
-			low: '368.07000',
-			close: '368.13000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:48:00',
-			open: '368.16000',
-			high: '368.16000',
-			low: '368.13010',
-			close: '368.15000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:47:00',
-			open: '368.10000',
-			high: '369.30000',
-			low: '368.10000',
-			close: '369.30000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:46:00',
-			open: '368.22000',
-			high: '368.22000',
-			low: '368.14000',
-			close: '368.14000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:45:00',
-			open: '368.26000',
-			high: '368.26000',
-			low: '368.22000',
-			close: '368.22000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:44:00',
-			open: '368.21000',
-			high: '368.22000',
-			low: '368.15000',
-			close: '368.18000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:43:00',
-			open: '368.23000',
-			high: '368.23000',
-			low: '368.23000',
-			close: '368.23000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:42:00',
-			open: '368.25000',
-			high: '368.25000',
-			low: '368.19000',
-			close: '368.19000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:41:00',
-			open: '368.25000',
-			high: '368.27000',
-			low: '368.25000',
-			close: '368.25000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:40:00',
-			open: '368.14000',
-			high: '368.26000',
-			low: '368.14000',
-			close: '368.26000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:39:00',
-			open: '368.08000',
-			high: '368.09000',
-			low: '368.02000',
-			close: '368.09000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:38:00',
-			open: '368.08990',
-			high: '368.10000',
-			low: '368.04000',
-			close: '368.04000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:37:00',
-			open: '368.15000',
-			high: '368.15000',
-			low: '368.06000',
-			close: '368.08000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:36:00',
-			open: '368.19000',
-			high: '368.24000',
-			low: '368.19000',
-			close: '368.24000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:35:00',
-			open: '368.23000',
-			high: '368.24000',
-			low: '368.15500',
-			close: '368.20000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:34:00',
-			open: '368.17000',
-			high: '368.17000',
-			low: '368.17000',
-			close: '368.17000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:33:00',
-			open: '368.14000',
-			high: '368.17000',
-			low: '368.14000',
-			close: '368.17000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:32:00',
-			open: '368.12000',
-			high: '368.18000',
-			low: '368.12000',
-			close: '368.18000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:31:00',
-			open: '368.11200',
-			high: '368.16000',
-			low: '368.11000',
-			close: '368.12000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:30:00',
-			open: '368.09000',
-			high: '368.09000',
-			low: '368.09000',
-			close: '368.09000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:29:00',
-			open: '368.10000',
-			high: '368.10000',
-			low: '368.05000',
-			close: '368.05000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:28:00',
-			open: '368.12000',
-			high: '368.15000',
-			low: '368.10000',
-			close: '368.11000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:27:00',
-			open: '368.11000',
-			high: '368.11000',
-			low: '367.98000',
-			close: '368.11000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:26:00',
-			open: '368.07000',
-			high: '368.12000',
-			low: '368.04000',
-			close: '368.12000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:25:00',
-			open: '367.94000',
-			high: '368.07000',
-			low: '367.94000',
-			close: '368.06000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:24:00',
-			open: '368.28000',
-			high: '368.28000',
-			low: '367.95000',
-			close: '367.95000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:23:00',
-			open: '368.32000',
-			high: '368.34000',
-			low: '368.29000',
-			close: '368.29000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:22:00',
-			open: '368.29000',
-			high: '368.29000',
-			low: '368.28000',
-			close: '368.28000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:21:00',
-			open: '368.28000',
-			high: '368.32000',
-			low: '368.28000',
-			close: '368.30900',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:20:00',
-			open: '368.27000',
-			high: '368.30000',
-			low: '368.23000',
-			close: '368.29000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:19:00',
-			open: '368.30000',
-			high: '368.36000',
-			low: '368.30000',
-			close: '368.30000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:18:00',
-			open: '368.34000',
-			high: '368.34000',
-			low: '368.27000',
-			close: '368.27000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:17:00',
-			open: '368.35000',
-			high: '368.35000',
-			low: '368.30000',
-			close: '368.35000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:16:00',
-			open: '368.42000',
-			high: '368.42000',
-			low: '368.28000',
-			close: '368.34000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:15:00',
-			open: '368.52000',
-			high: '368.52000',
-			low: '368.40000',
-			close: '368.40000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:14:00',
-			open: '368.55000',
-			high: '368.55000',
-			low: '368.55000',
-			close: '368.55000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:13:00',
-			open: '368.65000',
-			high: '368.65000',
-			low: '368.61000',
-			close: '368.61000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:12:00',
-			open: '368.63010',
-			high: '368.67000',
-			low: '368.63010',
-			close: '368.67000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:11:00',
-			open: '368.55000',
-			high: '368.64000',
-			low: '368.55000',
-			close: '368.63000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:10:00',
-			open: '368.56000',
-			high: '368.56000',
-			low: '368.51000',
-			close: '368.55000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:09:00',
-			open: '368.61000',
-			high: '368.61000',
-			low: '368.57000',
-			close: '368.57000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:08:00',
-			open: '368.55000',
-			high: '368.57000',
-			low: '368.54000',
-			close: '368.54000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:07:00',
-			open: '368.58000',
-			high: '368.60000',
-			low: '368.56000',
-			close: '368.56000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:06:00',
-			open: '368.69000',
-			high: '368.69000',
-			low: '368.58000',
-			close: '368.59000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:05:00',
-			open: '368.86000',
-			high: '368.86000',
-			low: '368.69000',
-			close: '368.69000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:04:00',
-			open: '368.81000',
-			high: '368.84000',
-			low: '368.80000',
-			close: '368.84000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:03:00',
-			open: '368.74000',
-			high: '368.77000',
-			low: '368.68000',
-			close: '368.77000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:02:00',
-			open: '368.68000',
-			high: '368.84000',
-			low: '368.68000',
-			close: '368.73000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:00:00',
-			open: '368.36010',
-			high: '368.77000',
-			low: '368.10000',
-			close: '368.54000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:59:00',
-			open: '368.66000',
-			high: '368.66000',
-			low: '368.64000',
-			close: '368.66000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:58:00',
-			open: '368.66000',
-			high: '368.66000',
-			low: '368.64000',
-			close: '368.64000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:57:00',
-			open: '368.68000',
-			high: '368.68000',
-			low: '368.66000',
-			close: '368.66000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:56:00',
-			open: '368.66000',
-			high: '368.70000',
-			low: '368.66000',
-			close: '368.69000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:55:00',
-			open: '368.63000',
-			high: '368.63000',
-			low: '368.63000',
-			close: '368.63000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:54:00',
-			open: '368.67000',
-			high: '368.68000',
-			low: '368.67000',
-			close: '368.68000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:53:00',
-			open: '368.65000',
-			high: '368.66000',
-			low: '368.63000',
-			close: '368.66000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:52:00',
-			open: '368.72000',
-			high: '368.72000',
-			low: '368.67000',
-			close: '368.67000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:51:00',
-			open: '368.71000',
-			high: '368.73000',
-			low: '368.71000',
-			close: '368.73000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:50:00',
-			open: '368.74000',
-			high: '368.76000',
-			low: '368.71000',
-			close: '368.72000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:49:00',
-			open: '368.75000',
-			high: '368.77000',
-			low: '368.75000',
-			close: '368.76000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:48:00',
-			open: '368.72000',
-			high: '368.75000',
-			low: '368.72000',
-			close: '368.72900',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:47:00',
-			open: '368.74000',
-			high: '368.74000',
-			low: '368.74000',
-			close: '368.74000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:46:00',
-			open: '368.73000',
-			high: '368.73000',
-			low: '368.73000',
-			close: '368.73000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:45:00',
-			open: '368.75000',
-			high: '368.75000',
-			low: '368.75000',
-			close: '368.75000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:44:00',
-			open: '368.68000',
-			high: '368.76000',
-			low: '368.68000',
-			close: '368.76000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:43:00',
-			open: '368.72000',
-			high: '368.72000',
-			low: '368.71000',
-			close: '368.71000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:42:00',
-			open: '368.65000',
-			high: '368.74000',
-			low: '368.64000',
-			close: '368.74000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:41:00',
-			open: '368.63000',
-			high: '368.63000',
-			low: '368.63000',
-			close: '368.63000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:40:00',
-			open: '368.67000',
-			high: '368.69000',
-			low: '368.65000',
-			close: '368.65000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:39:00',
-			open: '368.65000',
-			high: '368.65000',
-			low: '368.60000',
-			close: '368.62000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:38:00',
-			open: '368.62000',
-			high: '368.62000',
-			low: '368.60000',
-			close: '368.60000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:37:00',
-			open: '368.58000',
-			high: '368.61000',
-			low: '368.55000',
-			close: '368.60000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:36:00',
-			open: '368.52000',
-			high: '368.61000',
-			low: '368.52000',
-			close: '368.61000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:35:00',
-			open: '368.42000',
-			high: '368.44000',
-			low: '368.42000',
-			close: '368.44000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:34:00',
-			open: '368.44000',
-			high: '368.44000',
-			low: '368.42000',
-			close: '368.42000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:33:00',
-			open: '368.45000',
-			high: '368.45000',
-			low: '368.45000',
-			close: '368.45000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:32:00',
-			open: '368.49000',
-			high: '368.49000',
-			low: '368.49000',
-			close: '368.49000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:31:00',
-			open: '368.49000',
-			high: '368.49000',
-			low: '368.48000',
-			close: '368.48000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:29:00',
-			open: '368.46000',
-			high: '368.47900',
-			low: '368.46000',
-			close: '368.47900',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:28:00',
-			open: '368.41000',
-			high: '368.43000',
-			low: '368.41000',
-			close: '368.43000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:27:00',
-			open: '368.41000',
-			high: '368.42000',
-			low: '368.40000',
-			close: '368.42000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:26:00',
-			open: '368.40000',
-			high: '368.40000',
-			low: '368.40000',
-			close: '368.40000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:25:00',
-			open: '368.35000',
-			high: '368.35000',
-			low: '368.35000',
-			close: '368.35000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:24:00',
-			open: '368.29000',
-			high: '368.30000',
-			low: '368.29000',
-			close: '368.30000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:22:00',
-			open: '368.36000',
-			high: '368.36000',
-			low: '368.30000',
-			close: '368.30000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:21:00',
-			open: '368.31000',
-			high: '368.36000',
-			low: '368.31000',
-			close: '368.36000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:19:00',
-			open: '368.24000',
-			high: '368.28000',
-			low: '368.24000',
-			close: '368.26000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:17:00',
-			open: '368.20000',
-			high: '368.20000',
-			low: '368.17000',
-			close: '368.17000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:15:00',
-			open: '368.20000',
-			high: '368.23000',
-			low: '368.20000',
-			close: '368.23000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:14:00',
-			open: '368.22000',
-			high: '368.22000',
-			low: '368.22000',
-			close: '368.22000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:13:00',
-			open: '368.15000',
-			high: '368.17000',
-			low: '368.10000',
-			close: '368.10000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:12:00',
-			open: '368.18000',
-			high: '368.18000',
-			low: '368.12000',
-			close: '368.12000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:11:00',
-			open: '368.20000',
-			high: '368.20000',
-			low: '368.20000',
-			close: '368.20000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:10:00',
-			open: '368.23000',
-			high: '368.23000',
-			low: '368.23000',
-			close: '368.23000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:09:00',
-			open: '368.25000',
-			high: '368.26000',
-			low: '368.24000',
-			close: '368.24000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:07:00',
-			open: '368.26000',
-			high: '368.29000',
-			low: '368.26000',
-			close: '368.26000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:06:00',
-			open: '368.44000',
-			high: '368.44000',
-			low: '368.42000',
-			close: '368.42000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:04:00',
-			open: '368.37000',
-			high: '368.45000',
-			low: '368.34000',
-			close: '368.43000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:03:00',
-			open: '368.35000',
-			high: '368.36000',
-			low: '368.35000',
-			close: '368.36000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:02:00',
-			open: '368.36000',
-			high: '368.39000',
-			low: '368.36000',
-			close: '368.38000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:00:00',
-			open: '368.51000',
-			high: '368.51000',
-			low: '368.48000',
-			close: '368.48000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:59:00',
-			open: '368.50000',
-			high: '368.50000',
-			low: '368.43000',
-			close: '368.43000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:58:00',
-			open: '368.44000',
-			high: '368.49000',
-			low: '368.44000',
-			close: '368.49000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:57:00',
-			open: '368.41000',
-			high: '368.41000',
-			low: '368.41000',
-			close: '368.41000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:56:00',
-			open: '368.44000',
-			high: '368.44000',
-			low: '368.44000',
-			close: '368.44000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:54:00',
-			open: '368.40000',
-			high: '368.45000',
-			low: '368.40000',
-			close: '368.43000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:53:00',
-			open: '368.41000',
-			high: '368.43000',
-			low: '368.40000',
-			close: '368.41000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:52:00',
-			open: '368.36000',
-			high: '368.47000',
-			low: '368.36000',
-			close: '368.41000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:51:00',
-			open: '368.25000',
-			high: '368.33000',
-			low: '368.25000',
-			close: '368.30000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:50:00',
-			open: '368.24000',
-			high: '368.25000',
-			low: '368.24000',
-			close: '368.25000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:48:00',
-			open: '368.07000',
-			high: '368.18000',
-			low: '368.07000',
-			close: '368.15000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:47:00',
-			open: '368.05000',
-			high: '368.05000',
-			low: '368.05000',
-			close: '368.05000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:46:00',
-			open: '368.03000',
-			high: '368.08000',
-			low: '368.02000',
-			close: '368.08000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:45:00',
-			open: '368.10000',
-			high: '368.10000',
-			low: '368.02000',
-			close: '368.02000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:42:00',
-			open: '368.12000',
-			high: '368.15000',
-			low: '368.12000',
-			close: '368.15000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:41:00',
-			open: '368.06000',
-			high: '368.06000',
-			low: '368.06000',
-			close: '368.06000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:40:00',
-			open: '368.00000',
-			high: '368.02000',
-			low: '368.00000',
-			close: '368.02000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:38:00',
-			open: '368.07000',
-			high: '368.12000',
-			low: '368.07000',
-			close: '368.12000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:37:00',
-			open: '368.01000',
-			high: '368.02000',
-			low: '368.00000',
-			close: '368.00000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:36:00',
-			open: '368.02000',
-			high: '368.02000',
-			low: '368.02000',
-			close: '368.02000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:35:00',
-			open: '368.04000',
-			high: '368.04000',
-			low: '368.04000',
-			close: '368.04000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:34:00',
-			open: '368.20000',
-			high: '368.20000',
-			low: '368.00000',
-			close: '368.05000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:33:00',
-			open: '368.20000',
-			high: '368.20000',
-			low: '368.17000',
-			close: '368.17000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:32:00',
-			open: '368.20000',
-			high: '368.20000',
-			low: '368.06000',
-			close: '368.07000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:31:00',
-			open: '368.30000',
-			high: '368.30000',
-			low: '368.24000',
-			close: '368.24000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:30:00',
-			open: '368.24000',
-			high: '368.24000',
-			low: '368.24000',
-			close: '368.24000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:29:00',
-			open: '368.25000',
-			high: '368.25000',
-			low: '368.20000',
-			close: '368.20000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:28:00',
-			open: '368.10000',
-			high: '368.12000',
-			low: '368.10000',
-			close: '368.12000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:27:00',
-			open: '368.13000',
-			high: '368.18000',
-			low: '368.13000',
-			close: '368.14000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:26:00',
-			open: '368.18000',
-			high: '368.18000',
-			low: '368.10000',
-			close: '368.10000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:25:00',
-			open: '368.16000',
-			high: '368.16000',
-			low: '368.16000',
-			close: '368.16000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:24:00',
-			open: '368.14000',
-			high: '368.14000',
-			low: '368.14000',
-			close: '368.14000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:23:00',
-			open: '368.18000',
-			high: '368.19000',
-			low: '368.18000',
-			close: '368.18000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:22:00',
-			open: '368.31000',
-			high: '368.31000',
-			low: '368.30000',
-			close: '368.30000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:21:00',
-			open: '368.28000',
-			high: '368.29000',
-			low: '368.24000',
-			close: '368.29000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:20:00',
-			open: '368.45000',
-			high: '368.45000',
-			low: '368.45000',
-			close: '368.45000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:18:00',
-			open: '368.46000',
-			high: '368.46000',
-			low: '368.41000',
-			close: '368.42000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:17:00',
-			open: '368.47000',
-			high: '368.47000',
-			low: '368.41000',
-			close: '368.41000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:16:00',
-			open: '368.42000',
-			high: '368.42000',
-			low: '368.42000',
-			close: '368.42000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:15:00',
-			open: '368.35000',
-			high: '368.41000',
-			low: '368.35000',
-			close: '368.41000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:13:00',
-			open: '368.29000',
-			high: '368.35000',
-			low: '368.29000',
-			close: '368.35000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:12:00',
-			open: '368.32000',
-			high: '368.34000',
-			low: '368.28000',
-			close: '368.29000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:11:00',
-			open: '368.30000',
-			high: '368.34000',
-			low: '368.30000',
-			close: '368.34000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:10:00',
-			open: '368.26000',
-			high: '368.35000',
-			low: '368.26000',
-			close: '368.35000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:09:00',
-			open: '368.54000',
-			high: '368.54000',
-			low: '368.21000',
-			close: '368.25000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:07:00',
-			open: '368.73000',
-			high: '368.81000',
-			low: '368.73000',
-			close: '368.81000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:03:00',
-			open: '368.78000',
-			high: '368.78000',
-			low: '368.78000',
-			close: '368.78000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:02:00',
-			open: '368.72000',
-			high: '368.74000',
-			low: '368.72000',
-			close: '368.74000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:00:00',
-			open: '368.52000',
-			high: '368.52000',
-			low: '368.52000',
-			close: '368.52000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:59:00',
-			open: '368.54000',
-			high: '368.59000',
-			low: '368.54000',
-			close: '368.59000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:58:00',
-			open: '368.56000',
-			high: '368.56000',
-			low: '368.56000',
-			close: '368.56000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:57:00',
-			open: '368.65000',
-			high: '368.65000',
-			low: '368.65000',
-			close: '368.65000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:56:00',
-			open: '368.59000',
-			high: '368.69000',
-			low: '368.59000',
-			close: '368.69000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:54:00',
-			open: '368.61000',
-			high: '368.61000',
-			low: '368.57000',
-			close: '368.59000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:53:00',
-			open: '368.65000',
-			high: '368.65000',
-			low: '368.60000',
-			close: '368.61000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:52:00',
-			open: '368.65000',
-			high: '368.65000',
-			low: '368.65000',
-			close: '368.65000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:51:00',
-			open: '368.57000',
-			high: '368.63000',
-			low: '368.57000',
-			close: '368.63000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:50:00',
-			open: '368.50000',
-			high: '368.56000',
-			low: '368.43000',
-			close: '368.56000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:49:00',
-			open: '368.55000',
-			high: '368.55000',
-			low: '368.47000',
-			close: '368.47000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:48:00',
-			open: '368.58000',
-			high: '368.58000',
-			low: '368.56000',
-			close: '368.56000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:47:00',
-			open: '368.64000',
-			high: '368.64000',
-			low: '368.58000',
-			close: '368.58000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:46:00',
-			open: '368.58000',
-			high: '368.58000',
-			low: '368.56000',
-			close: '368.56000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:45:00',
-			open: '368.73000',
-			high: '368.73000',
-			low: '368.50000',
-			close: '368.50000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:44:00',
-			open: '368.76000',
-			high: '368.76000',
-			low: '368.76000',
-			close: '368.76000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:43:00',
-			open: '368.76000',
-			high: '368.76000',
-			low: '368.76000',
-			close: '368.76000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:42:00',
-			open: '368.78000',
-			high: '368.78000',
-			low: '368.72000',
-			close: '368.74000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:41:00',
-			open: '368.80000',
-			high: '368.80000',
-			low: '368.68000',
-			close: '368.68000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:40:00',
-			open: '368.78000',
-			high: '368.81000',
-			low: '368.77000',
-			close: '368.77000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:39:00',
-			open: '368.75000',
-			high: '368.85000',
-			low: '368.75000',
-			close: '368.77000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:38:00',
-			open: '368.90000',
-			high: '368.90000',
-			low: '368.77000',
-			close: '368.77000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:37:00',
-			open: '369.00000',
-			high: '369.00000',
-			low: '369.00000',
-			close: '369.00000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:33:00',
-			open: '369.01000',
-			high: '369.01000',
-			low: '369.00000',
-			close: '369.00000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:32:00',
-			open: '369.04000',
-			high: '369.04000',
-			low: '369.01000',
-			close: '369.01000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:30:00',
-			open: '369.03000',
-			high: '369.05000',
-			low: '369.03000',
-			close: '369.05000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:29:00',
-			open: '368.96000',
-			high: '368.96000',
-			low: '368.96000',
-			close: '368.96000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:28:00',
-			open: '368.91000',
-			high: '368.93000',
-			low: '368.90000',
-			close: '368.93000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:27:00',
-			open: '369.00000',
-			high: '369.00000',
-			low: '368.95000',
-			close: '368.95000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:26:00',
-			open: '369.00000',
-			high: '369.00000',
-			low: '369.00000',
-			close: '369.00000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:25:00',
-			open: '369.05000',
-			high: '369.05000',
-			low: '369.01000',
-			close: '369.05000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:24:00',
-			open: '369.08000',
-			high: '369.08000',
-			low: '369.03000',
-			close: '369.08000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:23:00',
-			open: '369.18000',
-			high: '369.18000',
-			low: '369.14000',
-			close: '369.14000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:19:00',
-			open: '369.18000',
-			high: '369.18000',
-			low: '369.18000',
-			close: '369.18000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:18:00',
-			open: '369.23000',
-			high: '369.23000',
-			low: '369.23000',
-			close: '369.23000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:17:00',
-			open: '369.18000',
-			high: '369.18000',
-			low: '369.18000',
-			close: '369.18000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:16:00',
-			open: '369.23000',
-			high: '369.23000',
-			low: '369.23000',
-			close: '369.23000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:15:00',
-			open: '369.27000',
-			high: '369.27000',
-			low: '369.27000',
-			close: '369.27000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:14:00',
-			open: '369.23000',
-			high: '369.23000',
-			low: '369.22000',
-			close: '369.22000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:13:00',
-			open: '369.30000',
-			high: '369.30000',
-			low: '369.23000',
-			close: '369.23000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:12:00',
-			open: '369.28000',
-			high: '369.28000',
-			low: '369.25000',
-			close: '369.25000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:11:00',
-			open: '369.27000',
-			high: '369.27000',
-			low: '369.27000',
-			close: '369.27000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:09:00',
-			open: '369.25000',
-			high: '369.25000',
-			low: '369.25000',
-			close: '369.25000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:08:00',
-			open: '369.25000',
-			high: '369.28000',
-			low: '369.25000',
-			close: '369.28000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:07:00',
-			open: '369.21000',
-			high: '369.24000',
-			low: '369.21000',
-			close: '369.24000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:04:00',
-			open: '369.20000',
-			high: '369.25000',
-			low: '369.20000',
-			close: '369.25000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:02:00',
-			open: '369.21000',
-			high: '369.24000',
-			low: '369.21000',
-			close: '369.24000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:00:00',
-			open: '369.23000',
-			high: '369.23000',
-			low: '369.21000',
-			close: '369.21000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:59:00',
-			open: '369.22000',
-			high: '369.22000',
-			low: '369.21000',
-			close: '369.21000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:58:00',
-			open: '369.17000',
-			high: '369.20000',
-			low: '369.17000',
-			close: '369.20000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:57:00',
-			open: '369.26000',
-			high: '369.26000',
-			low: '369.26000',
-			close: '369.26000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:56:00',
-			open: '369.34000',
-			high: '369.34000',
-			low: '369.34000',
-			close: '369.34000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:55:00',
-			open: '369.33000',
-			high: '369.38000',
-			low: '369.30000',
-			close: '369.38000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:54:00',
-			open: '369.37000',
-			high: '369.37000',
-			low: '369.30000',
-			close: '369.30000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:52:00',
-			open: '369.38000',
-			high: '369.38000',
-			low: '369.38000',
-			close: '369.38000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:51:00',
-			open: '369.47000',
-			high: '369.47000',
-			low: '369.45000',
-			close: '369.45000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:50:00',
-			open: '369.35000',
-			high: '369.42000',
-			low: '369.34000',
-			close: '369.42000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:49:00',
-			open: '369.46000',
-			high: '369.46000',
-			low: '369.43000',
-			close: '369.43000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:48:00',
-			open: '369.49000',
-			high: '369.49000',
-			low: '369.44000',
-			close: '369.45000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:47:00',
-			open: '369.54000',
-			high: '369.56000',
-			low: '369.54000',
-			close: '369.56000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:46:00',
-			open: '369.51000',
-			high: '369.53000',
-			low: '369.51000',
-			close: '369.53000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:45:00',
-			open: '369.55000',
-			high: '369.55000',
-			low: '369.50000',
-			close: '369.50000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:43:00',
-			open: '369.60000',
-			high: '369.60000',
-			low: '369.60000',
-			close: '369.60000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:42:00',
-			open: '369.59000',
-			high: '369.59000',
-			low: '369.59000',
-			close: '369.59000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:40:00',
-			open: '369.61000',
-			high: '369.61000',
-			low: '369.58000',
-			close: '369.58000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:38:00',
-			open: '369.71000',
-			high: '369.71000',
-			low: '369.71000',
-			close: '369.71000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:36:00',
-			open: '369.59000',
-			high: '369.59000',
-			low: '369.59000',
-			close: '369.59000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:33:00',
-			open: '369.62000',
-			high: '369.62000',
-			low: '369.51000',
-			close: '369.51000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:32:00',
-			open: '369.65000',
-			high: '369.65000',
-			low: '369.65000',
-			close: '369.65000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:31:00',
-			open: '369.60000',
-			high: '369.61000',
-			low: '369.60000',
-			close: '369.61000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:30:00',
-			open: '369.58000',
-			high: '369.58000',
-			low: '369.58000',
-			close: '369.58000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:29:00',
-			open: '369.56000',
-			high: '369.56000',
-			low: '369.50000',
-			close: '369.50000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:28:00',
-			open: '369.66000',
-			high: '369.66000',
-			low: '369.62000',
-			close: '369.62000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:27:00',
-			open: '369.80000',
-			high: '369.80000',
-			low: '369.71000',
-			close: '369.71000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:26:00',
-			open: '369.82000',
-			high: '369.82000',
-			low: '369.82000',
-			close: '369.82000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:25:00',
-			open: '369.85000',
-			high: '369.85000',
-			low: '369.85000',
-			close: '369.85000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:24:00',
-			open: '369.82000',
-			high: '369.82000',
-			low: '369.82000',
-			close: '369.82000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:22:00',
-			open: '369.83000',
-			high: '369.84000',
-			low: '369.83000',
-			close: '369.83000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:21:00',
-			open: '369.88000',
-			high: '369.88000',
-			low: '369.83000',
-			close: '369.83000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:19:00',
-			open: '369.83000',
-			high: '369.83000',
-			low: '369.75000',
-			close: '369.75000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:18:00',
-			open: '369.83000',
-			high: '369.83000',
-			low: '369.83000',
-			close: '369.83000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:17:00',
-			open: '369.83000',
-			high: '369.83000',
-			low: '369.83000',
-			close: '369.83000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:16:00',
-			open: '369.83000',
-			high: '369.83000',
-			low: '369.81000',
-			close: '369.83000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:14:00',
-			open: '369.84000',
-			high: '369.84000',
-			low: '369.84000',
-			close: '369.84000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:13:00',
-			open: '369.83000',
-			high: '369.83000',
-			low: '369.83000',
-			close: '369.83000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:12:00',
-			open: '369.69000',
-			high: '369.80000',
-			low: '369.69000',
-			close: '369.80000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:11:00',
-			open: '369.72000',
-			high: '369.74000',
-			low: '369.70000',
-			close: '369.74000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:10:00',
-			open: '369.70000',
-			high: '369.70000',
-			low: '369.70000',
-			close: '369.70000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:09:00',
-			open: '369.76000',
-			high: '369.76000',
-			low: '369.76000',
-			close: '369.76000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:08:00',
-			open: '369.71000',
-			high: '369.75000',
-			low: '369.71000',
-			close: '369.73000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:07:00',
-			open: '369.71000',
-			high: '369.71000',
-			low: '369.71000',
-			close: '369.71000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:06:00',
-			open: '369.60000',
-			high: '369.65000',
-			low: '369.60000',
-			close: '369.65000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:05:00',
-			open: '369.60000',
-			high: '369.69000',
-			low: '369.59000',
-			close: '369.65000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:04:00',
-			open: '369.74000',
-			high: '369.74000',
-			low: '369.62000',
-			close: '369.62000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:03:00',
-			open: '369.67000',
-			high: '369.73000',
-			low: '369.67000',
-			close: '369.72000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:02:00',
-			open: '369.68000',
-			high: '369.75000',
-			low: '369.66000',
-			close: '369.67000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:00:00',
-			open: '369.74000',
-			high: '369.85000',
-			low: '369.74000',
-			close: '369.85000',
-			volume: '0'
-		}
-	],
-	status: 'ok'
-}
-
-export const ivm = {
-	meta: {
-		symbol: 'IWM',
-		interval: '1min',
-		currency: 'USD',
-		exchange_timezone: 'America/New_York',
-		exchange: 'NYSE',
-		type: 'ETF'
-	},
-	values: [
-		{
-			datetime: '2022-04-05 09:29:00',
-			open: '207.87000',
-			high: '207.90000',
-			low: '207.83000',
-			close: '207.90000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:27:00',
-			open: '208.03000',
-			high: '208.05000',
-			low: '208.03000',
-			close: '208.05000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:24:00',
-			open: '208.00000',
-			high: '208.00000',
-			low: '207.97000',
-			close: '207.97000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:23:00',
-			open: '207.99000',
-			high: '207.99000',
-			low: '207.99000',
-			close: '207.99000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:22:00',
-			open: '208.01000',
-			high: '208.01000',
-			low: '208.01000',
-			close: '208.01000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:21:00',
-			open: '208.02000',
-			high: '208.02000',
-			low: '208.02000',
-			close: '208.02000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:20:00',
-			open: '208.00000',
-			high: '208.00000',
-			low: '208.00000',
-			close: '208.00000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:18:00',
-			open: '208.02010',
-			high: '208.02010',
-			low: '208.02010',
-			close: '208.02010',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:17:00',
-			open: '208.02000',
-			high: '208.02000',
-			low: '208.01000',
-			close: '208.01000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:16:00',
-			open: '208.02000',
-			high: '208.02000',
-			low: '208.00000',
-			close: '208.00000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:15:00',
-			open: '207.97000',
-			high: '208.00000',
-			low: '207.97000',
-			close: '208.00000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:11:00',
-			open: '207.95000',
-			high: '207.95000',
-			low: '207.93000',
-			close: '207.93000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:10:00',
-			open: '207.98000',
-			high: '207.98000',
-			low: '207.93000',
-			close: '207.94000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:08:00',
-			open: '207.95000',
-			high: '207.98000',
-			low: '207.95000',
-			close: '207.98000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:07:00',
-			open: '207.91000',
-			high: '207.95000',
-			low: '207.91000',
-			close: '207.95000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:06:00',
-			open: '207.86000',
-			high: '207.88000',
-			low: '207.86000',
-			close: '207.88000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:05:00',
-			open: '207.92000',
-			high: '207.92000',
-			low: '207.92000',
-			close: '207.92000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:04:00',
-			open: '207.98000',
-			high: '207.98000',
-			low: '207.98000',
-			close: '207.98000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:03:00',
-			open: '207.94000',
-			high: '207.98000',
-			low: '207.94000',
-			close: '207.98000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:02:00',
-			open: '207.96000',
-			high: '207.97000',
-			low: '207.96000',
-			close: '207.97000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:01:00',
-			open: '207.93010',
-			high: '207.93010',
-			low: '207.93010',
-			close: '207.93010',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 09:00:00',
-			open: '207.87000',
-			high: '207.87000',
-			low: '207.87000',
-			close: '207.87000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:57:00',
-			open: '207.25470',
-			high: '207.83000',
-			low: '207.25470',
-			close: '207.83000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:56:00',
-			open: '207.91000',
-			high: '207.91000',
-			low: '207.85000',
-			close: '207.88000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:55:00',
-			open: '207.94000',
-			high: '207.94000',
-			low: '207.91000',
-			close: '207.91000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:54:00',
-			open: '207.92000',
-			high: '207.92000',
-			low: '207.91000',
-			close: '207.91000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:53:00',
-			open: '207.93000',
-			high: '207.93000',
-			low: '207.93000',
-			close: '207.93000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:52:00',
-			open: '207.98950',
-			high: '207.99000',
-			low: '207.97000',
-			close: '207.97000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:48:00',
-			open: '208.04000',
-			high: '208.04000',
-			low: '208.04000',
-			close: '208.04000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:47:00',
-			open: '208.03000',
-			high: '208.03000',
-			low: '208.03000',
-			close: '208.03000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:44:00',
-			open: '208.05000',
-			high: '208.05000',
-			low: '208.05000',
-			close: '208.05000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:40:00',
-			open: '208.05000',
-			high: '208.05000',
-			low: '208.05000',
-			close: '208.05000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:39:00',
-			open: '208.03000',
-			high: '208.03000',
-			low: '208.03000',
-			close: '208.03000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:37:00',
-			open: '207.99000',
-			high: '207.99000',
-			low: '207.99000',
-			close: '207.99000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:36:00',
-			open: '208.01000',
-			high: '208.01000',
-			low: '208.01000',
-			close: '208.01000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:35:00',
-			open: '207.98000',
-			high: '208.02000',
-			low: '207.97000',
-			close: '207.97000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:34:00',
-			open: '207.94000',
-			high: '207.94000',
-			low: '207.93000',
-			close: '207.93000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:33:00',
-			open: '207.96000',
-			high: '207.97000',
-			low: '207.96000',
-			close: '207.96000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:32:00',
-			open: '207.96000',
-			high: '207.96000',
-			low: '207.94000',
-			close: '207.96000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:31:00',
-			open: '207.97000',
-			high: '207.97000',
-			low: '207.92000',
-			close: '207.92000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:30:00',
-			open: '207.93000',
-			high: '207.93000',
-			low: '207.93000',
-			close: '207.93000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:29:00',
-			open: '207.92000',
-			high: '207.92000',
-			low: '207.88000',
-			close: '207.88000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:28:00',
-			open: '207.97000',
-			high: '208.00000',
-			low: '207.95000',
-			close: '207.95000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:27:00',
-			open: '207.95010',
-			high: '207.95010',
-			low: '207.92000',
-			close: '207.93000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:26:00',
-			open: '207.93000',
-			high: '207.95000',
-			low: '207.93000',
-			close: '207.95000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:25:00',
-			open: '207.93000',
-			high: '207.94000',
-			low: '207.91000',
-			close: '207.91000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:24:00',
-			open: '208.06000',
-			high: '208.06000',
-			low: '207.97000',
-			close: '207.97000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:23:00',
-			open: '208.07000',
-			high: '208.07000',
-			low: '208.07000',
-			close: '208.07000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:20:00',
-			open: '207.97000',
-			high: '207.97000',
-			low: '207.97000',
-			close: '207.97000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:19:00',
-			open: '207.99000',
-			high: '207.99000',
-			low: '207.99000',
-			close: '207.99000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:18:00',
-			open: '208.00000',
-			high: '208.00000',
-			low: '208.00000',
-			close: '208.00000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:16:00',
-			open: '207.92000',
-			high: '207.92000',
-			low: '207.92000',
-			close: '207.92000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:15:00',
-			open: '207.97000',
-			high: '207.97000',
-			low: '207.95000',
-			close: '207.95000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:14:00',
-			open: '207.99000',
-			high: '207.99000',
-			low: '207.99000',
-			close: '207.99000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:11:00',
-			open: '207.98000',
-			high: '207.98000',
-			low: '207.98000',
-			close: '207.98000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:08:00',
-			open: '208.01000',
-			high: '208.03000',
-			low: '208.01000',
-			close: '208.03000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:07:00',
-			open: '208.00000',
-			high: '208.00000',
-			low: '208.00000',
-			close: '208.00000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:04:00',
-			open: '208.12000',
-			high: '208.12000',
-			low: '208.12000',
-			close: '208.12000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:03:00',
-			open: '208.08000',
-			high: '208.08000',
-			low: '208.08000',
-			close: '208.08000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:02:00',
-			open: '208.05000',
-			high: '208.09000',
-			low: '208.05000',
-			close: '208.09000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:01:00',
-			open: '208.05010',
-			high: '208.07000',
-			low: '208.05010',
-			close: '208.07000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 08:00:00',
-			open: '207.92990',
-			high: '208.01000',
-			low: '207.74000',
-			close: '207.97000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:59:00',
-			open: '208.02000',
-			high: '208.02000',
-			low: '208.02000',
-			close: '208.02000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:57:00',
-			open: '207.98000',
-			high: '208.00000',
-			low: '207.98000',
-			close: '208.00000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:52:00',
-			open: '208.02000',
-			high: '208.02000',
-			low: '208.02000',
-			close: '208.02000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:51:00',
-			open: '208.03000',
-			high: '208.03000',
-			low: '208.03000',
-			close: '208.03000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:50:00',
-			open: '208.02000',
-			high: '208.02000',
-			low: '208.02000',
-			close: '208.02000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:49:00',
-			open: '208.01000',
-			high: '208.06000',
-			low: '208.01000',
-			close: '208.06000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:44:00',
-			open: '207.98000',
-			high: '207.98000',
-			low: '207.98000',
-			close: '207.98000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:43:00',
-			open: '207.99000',
-			high: '208.03000',
-			low: '207.99000',
-			close: '207.99000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:42:00',
-			open: '207.95000',
-			high: '207.97000',
-			low: '207.95000',
-			close: '207.97000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:41:00',
-			open: '207.92000',
-			high: '207.92000',
-			low: '207.92000',
-			close: '207.92000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:40:00',
-			open: '207.94000',
-			high: '207.98000',
-			low: '207.94000',
-			close: '207.98000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:39:00',
-			open: '207.92000',
-			high: '207.92000',
-			low: '207.92000',
-			close: '207.92000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:38:00',
-			open: '207.91000',
-			high: '207.91000',
-			low: '207.87000',
-			close: '207.87000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:37:00',
-			open: '207.91000',
-			high: '207.91000',
-			low: '207.90000',
-			close: '207.91000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:36:00',
-			open: '207.94000',
-			high: '207.94000',
-			low: '207.91000',
-			close: '207.91000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:34:00',
-			open: '207.89000',
-			high: '207.89000',
-			low: '207.86000',
-			close: '207.86000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:32:00',
-			open: '207.94000',
-			high: '207.95000',
-			low: '207.94000',
-			close: '207.95000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:31:00',
-			open: '207.91000',
-			high: '207.91000',
-			low: '207.88000',
-			close: '207.89000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:30:00',
-			open: '207.91000',
-			high: '207.91000',
-			low: '207.91000',
-			close: '207.91000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:29:00',
-			open: '207.90000',
-			high: '207.90000',
-			low: '207.90000',
-			close: '207.90000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:27:00',
-			open: '207.87000',
-			high: '207.89000',
-			low: '207.87000',
-			close: '207.89000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:26:00',
-			open: '207.85000',
-			high: '207.85000',
-			low: '207.85000',
-			close: '207.85000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:25:00',
-			open: '207.78000',
-			high: '207.81000',
-			low: '207.78000',
-			close: '207.81000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:23:00',
-			open: '207.78000',
-			high: '207.78000',
-			low: '207.78000',
-			close: '207.78000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:22:00',
-			open: '207.87000',
-			high: '207.87000',
-			low: '207.82000',
-			close: '207.82000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:21:00',
-			open: '207.85000',
-			high: '207.85000',
-			low: '207.85000',
-			close: '207.85000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:19:00',
-			open: '207.79000',
-			high: '207.79000',
-			low: '207.79000',
-			close: '207.79000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:17:00',
-			open: '207.78000',
-			high: '207.79000',
-			low: '207.78000',
-			close: '207.79000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:16:00',
-			open: '207.78000',
-			high: '207.78000',
-			low: '207.76000',
-			close: '207.76000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:15:00',
-			open: '207.78000',
-			high: '207.78000',
-			low: '207.75000',
-			close: '207.75000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:14:00',
-			open: '207.71000',
-			high: '207.71000',
-			low: '207.71000',
-			close: '207.71000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:13:00',
-			open: '207.72000',
-			high: '207.72000',
-			low: '207.72000',
-			close: '207.72000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:12:00',
-			open: '207.72000',
-			high: '207.72000',
-			low: '207.67000',
-			close: '207.67000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:11:00',
-			open: '207.73000',
-			high: '207.73000',
-			low: '207.73000',
-			close: '207.73000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:09:00',
-			open: '207.71000',
-			high: '207.71000',
-			low: '207.71000',
-			close: '207.71000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:07:00',
-			open: '207.71000',
-			high: '207.71000',
-			low: '207.71000',
-			close: '207.71000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:06:00',
-			open: '207.78000',
-			high: '207.80000',
-			low: '207.74000',
-			close: '207.80000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:05:00',
-			open: '207.83000',
-			high: '207.83000',
-			low: '207.83000',
-			close: '207.83000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:04:00',
-			open: '207.79000',
-			high: '207.79000',
-			low: '207.79000',
-			close: '207.79000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:02:00',
-			open: '207.81000',
-			high: '207.81000',
-			low: '207.81000',
-			close: '207.81000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:01:00',
-			open: '207.81000',
-			high: '207.82000',
-			low: '207.81000',
-			close: '207.82000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 07:00:00',
-			open: '207.74000',
-			high: '207.74000',
-			low: '207.74000',
-			close: '207.74000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:58:00',
-			open: '207.76000',
-			high: '207.83000',
-			low: '207.76000',
-			close: '207.83000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:55:00',
-			open: '207.77000',
-			high: '207.77000',
-			low: '207.77000',
-			close: '207.77000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:54:00',
-			open: '207.80000',
-			high: '207.80000',
-			low: '207.80000',
-			close: '207.80000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:50:00',
-			open: '207.68000',
-			high: '207.68000',
-			low: '207.68000',
-			close: '207.68000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:49:00',
-			open: '207.68000',
-			high: '207.68000',
-			low: '207.68000',
-			close: '207.68000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:45:00',
-			open: '207.65000',
-			high: '207.65000',
-			low: '207.65000',
-			close: '207.65000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:44:00',
-			open: '207.64000',
-			high: '207.64000',
-			low: '207.61000',
-			close: '207.61000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:43:00',
-			open: '207.71000',
-			high: '207.71000',
-			low: '207.71000',
-			close: '207.71000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:42:00',
-			open: '207.65000',
-			high: '207.65000',
-			low: '207.65000',
-			close: '207.65000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:36:00',
-			open: '207.47000',
-			high: '207.47000',
-			low: '207.47000',
-			close: '207.47000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:35:00',
-			open: '207.54000',
-			high: '207.54000',
-			low: '207.46000',
-			close: '207.46000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:34:00',
-			open: '207.59000',
-			high: '207.59000',
-			low: '207.59000',
-			close: '207.59000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:33:00',
-			open: '207.56000',
-			high: '207.56000',
-			low: '207.56000',
-			close: '207.56000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:32:00',
-			open: '207.66000',
-			high: '207.66000',
-			low: '207.66000',
-			close: '207.66000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:30:00',
-			open: '207.58000',
-			high: '207.58000',
-			low: '207.58000',
-			close: '207.58000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:29:00',
-			open: '207.56000',
-			high: '207.58000',
-			low: '207.56000',
-			close: '207.58000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:28:00',
-			open: '207.59000',
-			high: '207.59000',
-			low: '207.59000',
-			close: '207.59000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:25:00',
-			open: '207.54000',
-			high: '207.55000',
-			low: '207.54000',
-			close: '207.55000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:24:00',
-			open: '207.51000',
-			high: '207.51000',
-			low: '207.51000',
-			close: '207.51000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:23:00',
-			open: '207.52000',
-			high: '207.52000',
-			low: '207.52000',
-			close: '207.52000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:22:00',
-			open: '207.60000',
-			high: '207.60000',
-			low: '207.58000',
-			close: '207.60000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:21:00',
-			open: '207.68000',
-			high: '207.68000',
-			low: '207.60000',
-			close: '207.60000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:20:00',
-			open: '207.70000',
-			high: '207.70000',
-			low: '207.70000',
-			close: '207.70000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:17:00',
-			open: '207.64000',
-			high: '207.64000',
-			low: '207.64000',
-			close: '207.64000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:14:00',
-			open: '207.53000',
-			high: '207.57000',
-			low: '207.53000',
-			close: '207.57000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:13:00',
-			open: '207.45000',
-			high: '207.46000',
-			low: '207.45000',
-			close: '207.46000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:12:00',
-			open: '207.52000',
-			high: '207.52000',
-			low: '207.48000',
-			close: '207.48000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:11:00',
-			open: '207.60000',
-			high: '207.60000',
-			low: '207.60000',
-			close: '207.60000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:10:00',
-			open: '207.55000',
-			high: '207.60000',
-			low: '207.55000',
-			close: '207.60000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:09:00',
-			open: '207.76000',
-			high: '207.76000',
-			low: '207.56000',
-			close: '207.61000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:08:00',
-			open: '207.78000',
-			high: '207.78000',
-			low: '207.78000',
-			close: '207.78000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:05:00',
-			open: '207.76000',
-			high: '207.76000',
-			low: '207.76000',
-			close: '207.76000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:04:00',
-			open: '207.80000',
-			high: '207.80000',
-			low: '207.80000',
-			close: '207.80000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:03:00',
-			open: '207.77000',
-			high: '207.77000',
-			low: '207.77000',
-			close: '207.77000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 06:00:00',
-			open: '207.66000',
-			high: '207.66000',
-			low: '207.66000',
-			close: '207.66000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:57:00',
-			open: '207.71000',
-			high: '207.72000',
-			low: '207.71000',
-			close: '207.72000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:53:00',
-			open: '207.81000',
-			high: '207.81000',
-			low: '207.81000',
-			close: '207.81000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:50:00',
-			open: '207.64000',
-			high: '207.65000',
-			low: '207.64000',
-			close: '207.65000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:49:00',
-			open: '207.69000',
-			high: '207.69000',
-			low: '207.69000',
-			close: '207.69000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:48:00',
-			open: '207.66000',
-			high: '207.66000',
-			low: '207.66000',
-			close: '207.66000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:43:00',
-			open: '207.87000',
-			high: '207.87000',
-			low: '207.85000',
-			close: '207.85000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:39:00',
-			open: '207.86000',
-			high: '207.86000',
-			low: '207.86000',
-			close: '207.86000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:38:00',
-			open: '207.91000',
-			high: '207.91000',
-			low: '207.90000',
-			close: '207.90000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:35:00',
-			open: '208.02000',
-			high: '208.03000',
-			low: '208.02000',
-			close: '208.03000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:29:00',
-			open: '207.95000',
-			high: '207.95000',
-			low: '207.95000',
-			close: '207.95000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:28:00',
-			open: '208.01000',
-			high: '208.01000',
-			low: '208.01000',
-			close: '208.01000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:27:00',
-			open: '208.04000',
-			high: '208.04000',
-			low: '208.01000',
-			close: '208.01000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:26:00',
-			open: '208.12000',
-			high: '208.12000',
-			low: '208.12000',
-			close: '208.12000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:25:00',
-			open: '208.08000',
-			high: '208.08000',
-			low: '208.06000',
-			close: '208.06000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:21:00',
-			open: '208.21000',
-			high: '208.21000',
-			low: '208.21000',
-			close: '208.21000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:17:00',
-			open: '208.11000',
-			high: '208.11000',
-			low: '208.11000',
-			close: '208.11000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:16:00',
-			open: '208.16000',
-			high: '208.16000',
-			low: '208.16000',
-			close: '208.16000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:14:00',
-			open: '208.19000',
-			high: '208.20000',
-			low: '208.19000',
-			close: '208.20000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:12:00',
-			open: '208.20000',
-			high: '208.22000',
-			low: '208.20000',
-			close: '208.22000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:11:00',
-			open: '208.21000',
-			high: '208.21000',
-			low: '208.20000',
-			close: '208.20000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:09:00',
-			open: '208.20000',
-			high: '208.20000',
-			low: '208.20000',
-			close: '208.20000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:07:00',
-			open: '208.18000',
-			high: '208.18000',
-			low: '208.18000',
-			close: '208.18000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:05:00',
-			open: '208.14000',
-			high: '208.17000',
-			low: '208.14000',
-			close: '208.17000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 05:02:00',
-			open: '208.17000',
-			high: '208.19000',
-			low: '208.14000',
-			close: '208.14000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:59:00',
-			open: '208.15000',
-			high: '208.15000',
-			low: '208.15000',
-			close: '208.15000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:58:00',
-			open: '208.05000',
-			high: '208.11000',
-			low: '208.05000',
-			close: '208.11000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:57:00',
-			open: '208.08000',
-			high: '208.08000',
-			low: '208.08000',
-			close: '208.08000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:56:00',
-			open: '208.20000',
-			high: '208.20000',
-			low: '208.20000',
-			close: '208.20000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:55:00',
-			open: '208.18000',
-			high: '208.18000',
-			low: '208.18000',
-			close: '208.18000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:53:00',
-			open: '208.22000',
-			high: '208.23000',
-			low: '208.22000',
-			close: '208.23000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:50:00',
-			open: '208.21000',
-			high: '208.21000',
-			low: '208.21000',
-			close: '208.21000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:49:00',
-			open: '208.20000',
-			high: '208.20000',
-			low: '208.20000',
-			close: '208.20000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:47:00',
-			open: '208.26000',
-			high: '208.27000',
-			low: '208.26000',
-			close: '208.27000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:45:00',
-			open: '208.21000',
-			high: '208.22000',
-			low: '208.21000',
-			close: '208.22000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:44:00',
-			open: '208.26000',
-			high: '208.26000',
-			low: '208.26000',
-			close: '208.26000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:42:00',
-			open: '208.25000',
-			high: '208.25000',
-			low: '208.25000',
-			close: '208.25000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:41:00',
-			open: '208.28000',
-			high: '208.28000',
-			low: '208.28000',
-			close: '208.28000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:38:00',
-			open: '208.24000',
-			high: '208.24000',
-			low: '208.24000',
-			close: '208.24000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:34:00',
-			open: '208.23000',
-			high: '208.23000',
-			low: '208.23000',
-			close: '208.23000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:30:00',
-			open: '208.23000',
-			high: '208.23000',
-			low: '208.23000',
-			close: '208.23000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:29:00',
-			open: '208.25000',
-			high: '208.25000',
-			low: '208.25000',
-			close: '208.25000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:28:00',
-			open: '208.32000',
-			high: '208.32000',
-			low: '208.32000',
-			close: '208.32000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:27:00',
-			open: '208.32000',
-			high: '208.32000',
-			low: '208.32000',
-			close: '208.32000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:25:00',
-			open: '208.35000',
-			high: '208.35000',
-			low: '208.35000',
-			close: '208.35000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:23:00',
-			open: '208.36000',
-			high: '208.36000',
-			low: '208.36000',
-			close: '208.36000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:22:00',
-			open: '208.42000',
-			high: '208.42000',
-			low: '208.38000',
-			close: '208.38000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:21:00',
-			open: '208.41000',
-			high: '208.44000',
-			low: '208.41000',
-			close: '208.42000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:20:00',
-			open: '208.45000',
-			high: '208.45000',
-			low: '208.45000',
-			close: '208.45000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:18:00',
-			open: '208.44000',
-			high: '208.44000',
-			low: '208.44000',
-			close: '208.44000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:16:00',
-			open: '208.47000',
-			high: '208.47000',
-			low: '208.46000',
-			close: '208.47000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:15:00',
-			open: '208.44000',
-			high: '208.49000',
-			low: '208.44000',
-			close: '208.48000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:14:00',
-			open: '208.44000',
-			high: '208.44000',
-			low: '208.43000',
-			close: '208.43000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:13:00',
-			open: '208.43000',
-			high: '208.43000',
-			low: '208.43000',
-			close: '208.43000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:12:00',
-			open: '208.33000',
-			high: '208.42000',
-			low: '208.33000',
-			close: '208.42000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:11:00',
-			open: '208.38000',
-			high: '208.38000',
-			low: '208.38000',
-			close: '208.38000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:09:00',
-			open: '208.39000',
-			high: '208.39000',
-			low: '208.39000',
-			close: '208.39000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:08:00',
-			open: '208.38000',
-			high: '208.38000',
-			low: '208.38000',
-			close: '208.38000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:07:00',
-			open: '208.36000',
-			high: '208.36000',
-			low: '208.31000',
-			close: '208.35000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:06:00',
-			open: '208.33000',
-			high: '208.35000',
-			low: '208.33000',
-			close: '208.35000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:05:00',
-			open: '208.33000',
-			high: '208.34000',
-			low: '208.33000',
-			close: '208.34000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:04:00',
-			open: '208.38000',
-			high: '208.38000',
-			low: '208.29000',
-			close: '208.29000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:03:00',
-			open: '208.36000',
-			high: '208.36000',
-			low: '208.35000',
-			close: '208.36000',
-			volume: '0'
-		},
-		{
-			datetime: '2022-04-05 04:02:00',
-			open: '208.26000',
-			high: '208.26000',
-			low: '208.26000',
-			close: '208.26000',
-			volume: '0'
-		}
-	],
-	status: 'ok'
-}
